requirements.txt
icecream==2.1.3
jsonlines==4.0.0
jsons==1.6.3
numpy==2.2.4
pandas==2.0.2
pyarrow==19.0.0
pyarrow==19.0.1
rbo==0.1.3
ruamel.base==1.0.0
toml==0.10.2
tqdm==4.65.0
tqdm==4.66.1
whalrus==0.4.6
